добавляет модули consts и utils
