+++ src/components/consts.js
@@ -1,0 +1,8 @@
+const EVENT_COUNT = 3;
+
+const Place = {
+  AFTERBEGIN: `afterbegin`,
+  BEFOREEND: `beforeend`
+};
+
+export {EVENT_COUNT, Place};
